refactor(nexus): refactored the dirs
