+++ crates/nexus/src/command/mod.rs
@@ -1,0 +1,4 @@
+pub mod aggregate;
+pub mod handler;
+pub mod pipeline;
+pub mod repository;
