+++ crates/nexus/src/lib.rs
@@ -1,9 +1,8 @@
 use serde::{Serialize, de::DeserializeOwned};
 use std::{error::Error, fmt::Debug, hash::Hash};
 
-pub mod aggregate;
-pub mod pipeline;
-pub mod repository;
+pub mod command;
+pub mod query;
 
 // Common marker trait for all message types in our system.
 pub trait Message: Debug + Send + Sync + 'static {}
